refactor: move utils file to utils folder
